+++ projects/syshub-rest-module/src/lib/svc/rest.service.ts
@@ -1427,11 +1427,33 @@ export class RestService {
    * true on success or an HttpErrorResponse in case of any error.
    */
   public login(username: string, password: string, keepLoggedin: boolean = true): BehaviorSubject<boolean | null | HttpErrorResponse> {
+    if (this.settings.useBasicAuth) {
+      if (this.settings.basic!.requiresLogin === false)
+        throw new Error('Method login not allowed for basic authentication');
+      return this.loginBasic(username, password, keepLoggedin);
+    } else {
+      return this.loginOauth(username, password, keepLoggedin);
+    }
+  }
+
+  private loginBasic(username: string, password: string, keepLoggedin: boolean = true): BehaviorSubject<boolean | null | HttpErrorResponse> {
+    this.session.setBasicToken({ username: username, password: password }, keepLoggedin);
+    let subject = new BehaviorSubject<boolean | null | HttpErrorResponse>(null);
+    this.getCurrentUser(true).subscribe((user) => {
+      if (user instanceof Error) {
+        this.session.clearToken();
+        subject.next(false);
+      }
+      else
+        subject.next(true);
+      subject.complete();
+    });
+    return subject;
+  }
+
+  private loginOauth(username: string, password: string, keepLoggedin: boolean = true): BehaviorSubject<boolean | null | HttpErrorResponse> {
     const serv = this;
     let subject = new BehaviorSubject<boolean | null | HttpErrorResponse>(null);
-    if (!this.settings.useOAuth) {
-      throw new Error('Method login not allowed for basic authentication')
-    }
     let body: string = `grant_type=password&username=${encodeURIComponent(username)}&password=${encodeURIComponent(password)}&`
       + `scope=${this.settings!.oauth!.scope}&client_id=${this.settings!.oauth!.clientId}&client_secret=${encodeURIComponent(this.settings!.oauth!.clientSecret!)}`;
     this.httpClient.post(`${this.settings!.host}webauth/oauth/token`, body, { observe: 'response' }).subscribe({
@@ -1446,11 +1468,11 @@ export class RestService {
             refreshToken: responseBody.refresh_token,
             username: username
           };
-          serv.session.setToken(token, keepLoggedin);
+          serv.session.setOauthToken(token, keepLoggedin);
           subject.next(true);
-          subject.complete();
         }
-        subject.next(false);
+        else
+          subject.next(false);
         subject.complete();
       },
       error: (e: HttpErrorResponse) => {
@@ -1702,7 +1724,7 @@ export class RestService {
           refreshToken: response.refresh_token,
           username: this.session.getUsername(),
         };
-        this.session.setToken(token);
+        this.session.setOauthToken(token);
         setTimeout(() => {
           // reset refreshing flag after a few millisecs so session has enough time to update anything
           this.isRefreshing$.next(false);